Crate gawk

Structs

Handler

Wrapper for code that handles Events of a specific type.

Publisher

Publishes all Events to all subscribed Handlers that accept Events of that type

Traits

DynEvent

Dynamically typed event. Used internally to alow Publishers to support Handlers and Events of multiple different types.

DynHandler

Dynamically typed Handler. Used internally to allow Publishers to support Events and Handlers of multiple different types.

Event

An object that a Publisher can send to its subscribers